requirements.txt
aenum==3.1.11
colorama==0.4.5
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
yapf==0.32.0
